Changed 400 to 502 in case of upstream error status code
Previously, this system gave a 400 to the frontend in case either the image server
or the further upstream stadsarchief server gave a non-404 error code (404s are
given as 404s). This is confusing and incorrect, because a 400 indicates a client
error, whereas the error could have any reason. I changed this to a 502, since
that is exactly wat it is; an error that this system gets from upstream in which
this server only acts as a gateway. The body of the message served to the
frontend will still contain the error code and message which this proxy/gateway
received from the upstream server.